xen-netback: add control grant mappings operations
Introduce support for staging grants to allow guests to have a
set of preallocated buffers that get reused over time. This is
handed over as control messages namely to get total mapping size,
add a list of grants and remove a list of grants.

These lists hand over a list of `struct xen_netif_gref_alloc` which
frontend provide (size of XEN_PAGE_SIZE which fits 512 entries). And
these entries contain the gref and flags to map into a backend
ballooned page, which gets added in a hash table of gref <-> backing
page on a per queue basis. Frontend can then use this to pregrant
certain pages and reuse them for Rx/Tx requests.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>