xen-netfront: generalize recycling for grants
Takes the already existent mechanism for recycling pages and leverages it
for grant references too. The difference though is that pages permanently
granted to the backend cannot be revoked (because those are mapped by the
other side) and hence these need to go to a separate quarantine pool, until
the point these pages can be consumed. The strategy is: 1) Get a page by
fetching oldest entry in rx_pool 2a) if it's not granted then the page
is freed at the head [or] 2b) if it's reusable return the page otherwise
if it's in use add it to quarantine pool and 3) try fetching oldest page
in quarantine pool and return if it's reusable.  Finally if all else
fails above above then we allocate a new page.  Worst case scenario if
we have two atomic read ops added on packet path in addition to
allocating a new page for Rx requests.

This page reuse strategy allows us to remove the grant copy for each page handed
over by the backend greatly improve guest RX performance to ~42-47 Gbit/s when
testing backend -> frontend. The measured recycling percentage is about
30% on TCP streams if pool size == ring size; and with pool size == 2 *
ring size these rises up to 80 - 100%.  This shows that bigger ring sizes
should allow for better recycling, which remains to be explored.

The downside of this approach is that it is not 100% guaranteed that
the Rx requests provided to the backend will be already mapped; in other
words, backend may need to do a grant copy on >=1% of the packets.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>