xen-netfront: copy to gref mappings for Tx buffers
This is done in order to allow user to avoid grant paths,
by resorting to copying instead. Since we aren't guaranteed
to reuse pages on stack transmit, hence we can only reliably
copy to the Tx granted pages. Hence we keep a shadow pool and
use that instead. We don't require coalescing Tx requests, as
we already linearize the skb when required slots > available
slots.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>