xen-netfront: copy to gref mappings for Tx buffers
This is done in order to allow user to avoid grant paths,
by resorting to copying instead on trasmit. Since we aren't guaranteed
to reuse pages on stack transmit, hence we can only reliably copy to the
Tx granted pages. Hence we keep a pregranted page per slot and copy to
that instead.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>