device-dax: Define a vm_operations_struct.access helper
Whem memory is not backed by a struct page but by a VM_SPECIAL vma,
get_user_pages() will fail. __access_remote_vm() handles that by calling
a vm operation, namely vm_ops.access(), which are meant to request the
device to copy from/to the PFN it was trying to access.

device-dax now supports PFN_SPECIAL, so we want to allow ptrace -- the
main user, alongside /proc/mem -- to be able to dump/access/change its
memory remotely.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>