mm/memory-failure: Handle failures for PFNs outside direct map
The handling of MCEs with PFNs associated with a pagemap means
the process gets a SIGBUS to handle the memory failure. For
memory which has no struct pages the kernel bails out and
does nothing.

With the introduction of pageless device-dax backed by hmem
we try to handle such failures, and thus retain the exact
same behaviour as those with a pagemap. If the process is
a VMM (e.g. QEMU) it is then responsible for injecting an
vMCE to the guest.

Introduce memory_failure_dev() for that purpose. It is quite
similar to memory_failure_dev_pagemap() except the way it collects
the processes mapping the PFN. Instead we consult the underlying
device ranges to see if it uses the failed PFN and add that
to the process list for MF_MUST_KILL, and derive the virtual
address based on the vm_start + dax_range pair.

Additionally we silence set_mce_nospec when called with a
!pfn_valid(pfn) given that we are not mapping those pfns in
the direct map.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>