device-dax: Add support for PFN_SPECIAL flags
Right now we assume there's gonna be a PFN_DEV|PFN_MAP which
means it will have a struct page backing the PFN but that is
not placed in normal system RAM zones.

Add support for PFN_DEV|PFN_SPECIAL only and therefore the
underlying vma won't have a struct page. For device dax, this
means not assuming callers will pass a dev_pagemap, introduce
the requirement for MADV_DONTFORK, finally not setting struct
page index/mapping on fault.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>