cover-letter: IOMMUFD Dirty Tracking
Presented herewith is a series that extends IOMMUFD to have IOMMU
hardware support for dirty bit in the IOPTEs.

Today, AMD Milan (which been out for a year now) supports it while ARM
SMMUv3.2 alongside VT-D rev3.x are expected to eventually come along.
The intended use-case is to support Live Migration with S{R-,}IOV, without
PCI devices having to supply its own IOVA dirty tracking hardware
blocks. And, of course hopefully[*] longterm it becomes how
IOVA dirty tracking is performed and relieve vendors to only care
about DMA quiescing/resume and device-state parts of VFIO migration.

[*] i.e. should it not greatly/drastically affect DMA performance and what not etc

At a quick glance, IOMMUFD lets the userspace create the IOAS with a
set of a IOVA ranges mapped to some physical memory composing an IO
pagetable. This is then attached to a particular device, consequently
creating the protection domain to share a common IO page table
representing the endporint DMA-addressable guest address space.
(Hopefully I am not twisting the terminology here) The resultant object
is an hw_pagetable object which represents the iommu_domain
object that will be directly manipulated. For more background on
IOMMUFD have a look at these two series[0][1] on the kernel and qemu
consumption respectivally. The IOMMUFD UAPI, kAPI and the iommu core
kAPI is then extended to provide:

1) Status and Toggling of Dirty Tracking on the iommu_domain. We model
as the most common case of changing hardware protection domain control
bits. The ARM specific case of having to enable the per-PTE DBM control
bit will be always enabled and setting/clearing tracking means just
clearing the dirty bits at start and at stop. The 'real' tracking of
whether dirty tracking is enabled or not is stored in the IOMMU driver,
hence no new fields are added to iommufd pagetable structures.

2) Read the I/O PTEs and marshal its dirtyiness into a bitmap. The bitmap
thus describe the IOVAs that got written by the device. While performing
the marshalling also drivers need to clear the dirty bits from IOPTE and
allow the kAPI caller to batch the much needed IOTLB flush.
There's no copy of bitmaps to userspace backed memory, all is zerocopy
based. So far this is a test-and-clear kind of interface given that the
IOPT walk is going to be expensive. It occured to me to separate
the readout of dirty, and the clearing of dirty from IOPTEs.
I haven't opted for that one, given that it would mean two lenghty IOPTE
walks and felt counter-performant.

The series is organized as follows:

* Patches 1-4: Takes care of the iommu domain operations to be added and
extends iommufd io-pagetable to set/clear dirty tracking, as well as
reading the dirty bits from the IOMMU pagetables. The idea is to abstract
iommu drivers from any idea of how bitmaps are stored or propagated back to
the caller, as well as allowing control/batching over IOTLB flush. So
there's a data structure and an helper that only tells the upper layer that
an IOVA range got dirty. This logic is shared with VFIO[2] and it's meant
to pin pages, walking the bitmap user memory, and kmap-ing them as
needed. IOMMU driver just has an idea of a 'dirty bitmap state' and
recording an IOVA as dirty.

* Patches 5-7: Adds the UAPIs for IOMMUFD, and selftests.
The selftests, test mainly the principal workflow, still needs
to get added more corner cases.

For completeness and most importantly to make sure the new IOMMU core
ops capture the hardware blocks, I've implemented for the IOMMUs that
will eventually get IOMMU A/D support. So the next half of the series
presents *proof of concept* implementations for IOMMUs:

* Patches 8-9: AMD IOMMU implementation, particularly on those having
HDSup support. Tested with a Qemu amd-iommu with HDSUp emulated,
and also on a AMD Milan server IOMMU.

* Patches 10: Intel IOMMU rev3.x implementation. Tested with a Qemu
based intel-iommu vIOMMU with SSADS/SLADS emulation support.

To help testing/prototypization, I also wrote qemu iommu emulation
to increase coverage of this code and hopefully make this more broadly
available for fellow contributors/devs. You can look it up here:

	https://github.com/jpemartins/qemu/commits/iommufd

Notable Remarks:

* Dirty bit tracking is not enough. Large IO pages tend to be the norm when
DMA mapping large ranges of IOVA space, when really the VMM wants the smallest
granuliarity possible (i.e. base pages). A separate bit of work will
need to take care demoting IOPTE page sizes at guest-runtime to
increase/decrease the dirty tracking granularity, likely under the form of
a IOAS demote/promote page-size within a previously mapped IOVA range.

Feedback is very much appreciated!

[0] https://lore.kernel.org/kvm/0-v1-e79cd8d168e8+6-iommufd_jgg@nvidia.com/
[1] https://lore.kernel.org/kvm/20220414104710.28534-1-yi.l.liu@intel.com/
[2] https://lore.kernel.org/kvm/20220705102740.29337-1-yishaih@nvidia.com/

Thanks,
	Joao

TODOs:
- Move into HWPT based iommu-domain flags instead of HWPT_GET_DIRTY [?]
- Move into faster walker function in iommu/intel
- More performance testing