iommu/amd: Add domain_alloc_user based domain allocation
Add the domain_alloc_user op implementation. To that end, refactor
amd_iommu_domain_alloc() to receive a dev pointer and flags, while
renaming it to .. such that it becomes a common function shared with
domain_alloc_user() implementation. The sole difference with
domain_alloc_user() is that we initialize also other fields that
iommu_domain_alloc() does. It lets it return the iommu domain
correctly initialized in one function.

This is in preparation to add dirty enforcement on AMD implementation
of domain_alloc_user.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>